chore: move backend to `api` folder
