+++ api/app.ts
@@ -1,5 +1,4 @@
 import express, { Request, RequestHandler, Response, Router } from 'express'
-import { version } from './package.json'
 import history from 'connect-history-api-fallback'
 import cors from 'cors'
 import csrf from 'csurf'
@@ -1158,7 +1157,7 @@ app.post(
 
 			// update versions to actual ones
 			settings.gateway.versions = {
-				app: version, // don't use getVersion here as it may include commit sha
+				app: utils.pkgJson.version, // don't use getVersion here as it may include commit sha
 				driver: libVersion,
 				server: serverVersion,
 			}
