+++ api/config/app.ts
@@ -10,15 +10,16 @@ export const logsDir: string = joinPath(storeDir, 'logs')
 export const snippetsDir: string = joinPath(storeDir, 'snippets')
 
 export const tmpDir: string = joinPath(storeDir, '.tmp')
-export const backupsDir: string = process.env.BACKUPS_DIR || joinPath(storeDir, 'backups')
+export const backupsDir: string =
+	process.env.BACKUPS_DIR || joinPath(storeDir, 'backups')
 export const nvmBackupsDir: string = joinPath(backupsDir, 'nvm')
 export const storeBackupsDir: string = joinPath(backupsDir, 'store')
 
-
-export const defaultUser: string =  'admin'
-export const defaultPsw: string =  'zwave'
+export const defaultUser: string = 'admin'
+export const defaultPsw: string = 'zwave'
 // lgtm [js/hardcoded-credentials]
-export const sessionSecret: string =  process.env.SESSION_SECRET || 'DEFAULT_SESSION_SECRET_CHANGE_ME'
-export const base: string =  process.env.BASE_PATH || '/'
-export const port: string | number =  process.env.PORT || 8091
-export const host: string =  process.env.HOST // by default undefined, so it will listen on all interfaces both ipv4 and ipv6
+export const sessionSecret: string =
+	process.env.SESSION_SECRET || 'DEFAULT_SESSION_SECRET_CHANGE_ME'
+export const base: string = process.env.BASE_PATH || '/'
+export const port: string | number = process.env.PORT || 8091
+export const host: string = process.env.HOST // by default undefined, so it will listen on all interfaces both ipv4 and ipv6
