+++ api/hass/configurations.ts
