+++ api/hass/devices.ts
