+++ api/lib/EventEmitter.ts
