+++ api/lib/Gateway.ts
@@ -1,10 +1,10 @@
 import * as fs from 'fs'
 import * as path from 'path'
-import * as utils from '../lib/utils'
+import * as utils from './utils'
 import { AlarmSensorType, SetValueAPIOptions } from 'zwave-js'
 import { CommandClasses, ValueID } from '@zwave-js/core'
 import * as Constants from './Constants'
-import { LogLevel, module } from '../lib/logger'
+import { LogLevel, module } from './logger'
 import hassCfg from '../hass/configurations'
 import hassDevices from '../hass/devices'
 import { storeDir } from '../config/app'
