+++ api/lib/SocketManager.ts
